Fixed bad behaviour of demo install when multiple instances are present
If `/usr/share/elasticsearch/bin/elasticsearch` is present then
`install_demo_configuration.sh` will always assume that it has been run from
a packaged version of elasticsearch. This may not be true when another instances
is started from the tar.gz package and the security plugin is installed there.

This commit makes the install demo script check if the BASE_DIR directory is
`/usr/share/elasticsearch` and only in this case it assumes that we are dealing
with a packaged installation.