Fix the security plugin so that we use X-Opaque-Id (#349)
The X-Opaque-Id header, when provided on the HTTP request header, would cause security plugin throws an exception. When a search request comes in, the security plugin stashes the context, hoping to start with a brand new, empty one, and re-insert headers in a whitelist (mostly security plugin headers). X-Opaque-Id is whitelisted too and re-inserted to the header. However, when we stash context, ES leaves out everything except X-Opaque-id. Thus, when we re-insert X-Opaque-id, we get exception indicating an X-Opaque-id is already present. This PR removes X-Opaque-Id from the security plugin's whitelisted headers.

This PR also fixes a test related to X-Opaque-Id: previously, it won't expose the bug, now it will.

Testing done:
- manually verified the fix mitigates the issue in a cluster.
- fix a test to expose this issue.